gadget,gadget/install: move partition creation to install
