+++ gadget/ondisk_test.go
@@ -20,13 +20,9 @@
 package gadget_test
 
 import (
-	"bytes"
-	"errors"
-	"fmt"
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"time"
 
 	. "gopkg.in/check.v1"
 
@@ -54,16 +50,7 @@ func (s *ondiskTestSuite) SetUpTest(c *C) {
 	c.Assert(err, IsNil)
 }
 
-const (
-	scriptPartitionsBios = iota
-	scriptPartitionsBiosSeed
-	scriptPartitionsBiosSeedData
-)
-
-func makeSfdiskScript(num int) string {
-	var b bytes.Buffer
-
-	b.WriteString(`
+const mockSfdiskScriptBiosSeed = `
 >&2 echo "Some warning from sfdisk"
 echo '{
   "partitiontable": {
@@ -73,11 +60,7 @@ echo '{
     "unit": "sectors",
     "firstlba": 34,
     "lastlba": 8388574,
-    "partitions": [`)
-
-	// BIOS boot partition
-	if num >= scriptPartitionsBios {
-		b.WriteString(`
+    "partitions": [
       {
         "node": "/dev/node1",
         "start": 2048,
@@ -85,12 +68,7 @@ echo '{
         "type": "21686148-6449-6E6F-744E-656564454649",
         "uuid": "2E59D969-52AB-430B-88AC-F83873519F6F",
         "name": "BIOS Boot"
-      }`)
-	}
-
-	// Seed partition
-	if num >= scriptPartitionsBiosSeed {
-		b.WriteString(`,
+      },
       {
         "node": "/dev/node2",
         "start": 4096,
@@ -99,62 +77,19 @@ echo '{
         "uuid": "44C3D5C3-CAE1-4306-83E8-DF437ACDB32F",
         "name": "Recovery",
         "attrs": "GUID:59"
-      }`)
-	}
-
-	// Data partition
-	if num >= scriptPartitionsBiosSeedData {
-		b.WriteString(`,
-      {
-        "node": "/dev/node3",
-        "start": 2461696,
-        "size": 2457600,
-        "type": "0FC63DAF-8483-4772-8E79-3D69D8477DE4",
-        "uuid": "f940029d-bfbb-4887-9d44-321e85c63866",
-        "name": "Writable",
-        "attrs": "GUID:59"
-      }`)
-	}
-
-	b.WriteString(`
+      }
     ]
   }
-}'`)
-	return b.String()
-}
+}'`
 
-func makeLsblkScript(num int) string {
-	var b bytes.Buffer
-
-	// BIOS boot partition
-	if num >= scriptPartitionsBios {
-		b.WriteString(`
+const mockLsblkScriptBiosSeed = `
 [ "$3" == "/dev/node1" ] && echo '{
     "blockdevices": [ {"name": "node1", "fstype": null, "label": null, "uuid": null, "mountpoint": null} ]
-}'`)
-	}
-
-	// Seed partition
-	if num >= scriptPartitionsBiosSeed {
-		b.WriteString(`
+}'
 [ "$3" == "/dev/node2" ] && echo '{
     "blockdevices": [ {"name": "node2", "fstype": "vfat", "label": "ubuntu-seed", "uuid": "A644-B807", "mountpoint": null} ]
-}'`)
-	}
-
-	// Data partition
-	if num >= scriptPartitionsBiosSeedData {
-		b.WriteString(`
-[ "$3" == "/dev/node3" ] && echo '{
-    "blockdevices": [ {"name": "node3", "fstype": "ext4", "label": "ubuntu-data", "uuid": "8781-433a", "mountpoint": null} ]
-}'`)
-	}
-
-	b.WriteString(`
-exit 0`)
-
-	return b.String()
-}
+}'
+exit 0`
 
 func makeMockGadget(gadgetRoot, gadgetContent string) error {
 	if err := os.MkdirAll(filepath.Join(gadgetRoot, "meta"), 0755); err != nil {
@@ -226,10 +161,10 @@ var mockOnDiskStructureWritable = gadget.OnDiskStructure{
 }
 
 func (s *ondiskTestSuite) TestDeviceInfoGPT(c *C) {
-	cmdSfdisk := testutil.MockCommand(c, "sfdisk", makeSfdiskScript(scriptPartitionsBiosSeed))
+	cmdSfdisk := testutil.MockCommand(c, "sfdisk", mockSfdiskScriptBiosSeed)
 	defer cmdSfdisk.Restore()
 
-	cmdLsblk := testutil.MockCommand(c, "lsblk", makeLsblkScript(scriptPartitionsBiosSeed))
+	cmdLsblk := testutil.MockCommand(c, "lsblk", mockLsblkScriptBiosSeed)
 	defer cmdLsblk.Restore()
 
 	dl, err := gadget.OnDiskVolumeFromDevice("/dev/node")
@@ -446,10 +381,10 @@ func (s *ondiskTestSuite) TestDeviceInfoError(c *C) {
 }
 
 func (s *ondiskTestSuite) TestBuildPartitionList(c *C) {
-	cmdSfdisk := testutil.MockCommand(c, "sfdisk", makeSfdiskScript(scriptPartitionsBiosSeed))
+	cmdSfdisk := testutil.MockCommand(c, "sfdisk", mockSfdiskScriptBiosSeed)
 	defer cmdSfdisk.Restore()
 
-	cmdLsblk := testutil.MockCommand(c, "lsblk", makeLsblkScript(scriptPartitionsBiosSeed))
+	cmdLsblk := testutil.MockCommand(c, "lsblk", mockLsblkScriptBiosSeed)
 	defer cmdLsblk.Restore()
 
 	ptable := gadget.SFDiskPartitionTable{
@@ -491,142 +426,6 @@ func (s *ondiskTestSuite) TestBuildPartitionList(c *C) {
 	c.Assert(sfdiskInput.String(), Equals, `/dev/node3 : start=     2461696, size=     5926879, type=0FC63DAF-8483-4772-8E79-3D69D8477DE4, name="Writable", attrs="GUID:59"
 `)
 	c.Assert(create, DeepEquals, []gadget.OnDiskStructure{mockOnDiskStructureWritable})
-}
-
-func (s *ondiskTestSuite) TestCreatePartitions(c *C) {
-	cmdSfdisk := testutil.MockCommand(c, "sfdisk", makeSfdiskScript(scriptPartitionsBiosSeed))
-	defer cmdSfdisk.Restore()
-
-	cmdLsblk := testutil.MockCommand(c, "lsblk", makeLsblkScript(scriptPartitionsBiosSeed))
-	defer cmdLsblk.Restore()
-
-	cmdPartx := testutil.MockCommand(c, "partx", "")
-	defer cmdPartx.Restore()
-
-	calls := 0
-	restore := gadget.MockEnsureNodesExist(func(ds []gadget.OnDiskStructure, timeout time.Duration) error {
-		calls++
-		c.Assert(ds, HasLen, 1)
-		c.Assert(ds[0].Node, Equals, "/dev/node3")
-		return nil
-	})
-	defer restore()
-
-	gadgetRoot := filepath.Join(c.MkDir(), "gadget")
-	err := makeMockGadget(gadgetRoot, gadgetContent)
-	c.Assert(err, IsNil)
-	pv, err := gadget.PositionedVolumeFromGadget(gadgetRoot)
-	c.Assert(err, IsNil)
-
-	dl, err := gadget.OnDiskVolumeFromDevice("/dev/node")
-	c.Assert(err, IsNil)
-	created, err := dl.CreateMissing(pv)
-	c.Assert(err, IsNil)
-	c.Assert(created, DeepEquals, []gadget.OnDiskStructure{mockOnDiskStructureWritable})
-	c.Assert(calls, Equals, 1)
-
-	// Check partition table read and write
-	c.Assert(cmdSfdisk.Calls(), DeepEquals, [][]string{
-		{"sfdisk", "--json", "-d", "/dev/node"},
-		{"sfdisk", "--append", "--no-reread", "/dev/node"},
-	})
-
-	// Check partition table update
-	c.Assert(cmdPartx.Calls(), DeepEquals, [][]string{
-		{"partx", "-u", "/dev/node"},
-	})
-}
-
-func (s *ondiskTestSuite) TestRemovePartitionsTrivial(c *C) {
-	// no locally created partitions
-	cmdSfdisk := testutil.MockCommand(c, "sfdisk", makeSfdiskScript(scriptPartitionsBios))
-	defer cmdSfdisk.Restore()
-
-	cmdLsblk := testutil.MockCommand(c, "lsblk", makeLsblkScript(scriptPartitionsBios))
-	defer cmdLsblk.Restore()
-
-	cmdPartx := testutil.MockCommand(c, "partx", "")
-	defer cmdPartx.Restore()
-
-	dl, err := gadget.OnDiskVolumeFromDevice("/dev/node")
-	c.Assert(err, IsNil)
-
-	err = dl.RemoveCreated()
-	c.Assert(err, IsNil)
-
-	c.Assert(cmdSfdisk.Calls(), DeepEquals, [][]string{
-		{"sfdisk", "--json", "-d", "/dev/node"},
-	})
-}
-
-func (s *ondiskTestSuite) TestRemovePartitions(c *C) {
-	const mockSfdiskScriptRemovablePartition = `
-if [ -f %[1]s/2 ]; then
-   rm %[1]s/[0-9]
-elif [ -f %[1]s/1 ]; then
-   touch %[1]s/2
-   exit 0
-else
-   PART=',{"node": "/dev/node2", "start": 4096, "size": 2457600, "type": "0FC63DAF-8483-4772-8E79-3D69D8477DE4", "uuid": "44C3D5C3-CAE1-4306-83E8-DF437ACDB32F", "name": "Recovery", "attrs": "GUID:59"}'
-   touch %[1]s/1
-fi
-echo '{
-   "partitiontable": {
-      "label": "gpt",
-      "id": "9151F25B-CDF0-48F1-9EDE-68CBD616E2CA",
-      "device": "/dev/node",
-      "unit": "sectors",
-      "firstlba": 34,
-      "lastlba": 8388574,
-      "partitions": [
-         {"node": "/dev/node1", "start": 2048, "size": 2048, "type": "21686148-6449-6E6F-744E-656564454649", "uuid": "2E59D969-52AB-430B-88AC-F83873519F6F", "name": "BIOS Boot"}
-         '"$PART
-      ]
-   }
-}"`
-
-	cmdSfdisk := testutil.MockCommand(c, "sfdisk", fmt.Sprintf(mockSfdiskScriptRemovablePartition, s.dir))
-	defer cmdSfdisk.Restore()
-
-	cmdLsblk := testutil.MockCommand(c, "lsblk", makeLsblkScript(scriptPartitionsBiosSeed))
-	defer cmdLsblk.Restore()
-
-	cmdPartx := testutil.MockCommand(c, "partx", "")
-	defer cmdPartx.Restore()
-
-	dl, err := gadget.OnDiskVolumeFromDevice("/dev/node")
-	c.Assert(err, IsNil)
-
-	c.Assert(cmdLsblk.Calls(), DeepEquals, [][]string{
-		{"lsblk", "--fs", "--json", "/dev/node1"},
-		{"lsblk", "--fs", "--json", "/dev/node2"},
-	})
-
-	err = dl.RemoveCreated()
-	c.Assert(err, IsNil)
-
-	c.Assert(cmdSfdisk.Calls(), DeepEquals, [][]string{
-		{"sfdisk", "--json", "-d", "/dev/node"},
-		{"sfdisk", "--no-reread", "--delete", "/dev/node", "2"},
-		{"sfdisk", "--json", "-d", "/dev/node"},
-	})
-}
-
-func (s *ondiskTestSuite) TestRemovePartitionsError(c *C) {
-	cmdSfdisk := testutil.MockCommand(c, "sfdisk", makeSfdiskScript(scriptPartitionsBiosSeedData))
-	defer cmdSfdisk.Restore()
-
-	cmdLsblk := testutil.MockCommand(c, "lsblk", makeLsblkScript(scriptPartitionsBiosSeedData))
-	defer cmdLsblk.Restore()
-
-	cmdPartx := testutil.MockCommand(c, "partx", "")
-	defer cmdPartx.Restore()
-
-	dl, err := gadget.OnDiskVolumeFromDevice("node")
-	c.Assert(err, IsNil)
-
-	err = dl.RemoveCreated()
-	c.Assert(err, ErrorMatches, "cannot remove partitions: /dev/node3")
 }
 
 func (s *ondiskTestSuite) TestListCreatedPartitionsGPT(c *C) {
@@ -802,51 +601,4 @@ func (s *ondiskTestSuite) TestFilesystemInfoError(c *C) {
 	info, err := gadget.FilesystemInfo("/dev/node")
 	c.Assert(err, ErrorMatches, "lsblk: not found")
 	c.Assert(info, IsNil)
-}
-
-func (s *ondiskTestSuite) TestEnsureNodesExist(c *C) {
-	for _, tc := range []struct {
-		utErr error
-		err   string
-	}{
-		{utErr: nil, err: ""},
-		{utErr: errors.New("some error"), err: "some error"},
-	} {
-		c.Logf("utErr:%v err:%q", tc.utErr, tc.err)
-
-		node := filepath.Join(c.MkDir(), "node")
-		err := ioutil.WriteFile(node, nil, 0644)
-		c.Assert(err, IsNil)
-
-		calls := 0
-		restore := gadget.MockInternalUdevTrigger(func(node string) error {
-			calls++
-			c.Assert(node, Equals, node)
-			return tc.utErr
-		})
-		defer restore()
-
-		ds := []gadget.OnDiskStructure{{Node: node}}
-		err = gadget.EnsureNodesExist(ds, 10*time.Millisecond)
-		if tc.err == "" {
-			c.Assert(err, IsNil)
-		} else {
-			c.Assert(err, ErrorMatches, tc.err)
-		}
-		c.Assert(calls, Equals, 1)
-	}
-}
-
-func (s *ondiskTestSuite) TestEnsureNodesExistTimeout(c *C) {
-	cmdUdevadm := testutil.MockCommand(c, "udevadm", "")
-	defer cmdUdevadm.Restore()
-
-	node := filepath.Join(c.MkDir(), "node")
-	ds := []gadget.OnDiskStructure{{Node: node}}
-	t := time.Now()
-	timeout := 1 * time.Second
-	err := gadget.EnsureNodesExist(ds, timeout)
-	c.Assert(err, ErrorMatches, fmt.Sprintf("device %s not available", node))
-	c.Assert(time.Since(t) >= timeout, Equals, true)
-	c.Assert(cmdUdevadm.Calls(), HasLen, 0)
 }
